Create a DoublyLinkedList struct with `insert` method
This commit introduces the struct DoublyLinkedList and a helper struct Node.
DoublyLinkedList is a container for many nodes, each of which has a link to
the previous and the next node in the list. The implementation uses Rc
module to keep track of shared pointers and RefCell to allow mutability of the nodes.
There is also one test written, testing the correctness of our `insert` method.